requirements.txt
fastapi==0.128.1
uvicorn==0.40.0
sqlalchemy==2.0.46
psycopg2-binary==2.9.11
google-genai==1.63.0
opencv-python-headless==4.10.0.84
numpy==2.2.6
pandas==2.3.3
openpyxl==3.1.5
pillow==12.1.0
python-multipart==0.0.22
fpdf2==2.8.7
pydantic-settings==2.12.0
python-dotenv==1.2.1
pyseto==1.9.1
passlib[argon2]==1.7.4
protobuf==5.29.3
jinja2==3.1.5